Añadido sweetAlert al eliminar artículos

+++ AprendiendoAngular/src/app/components/article/article.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
 import { ArticleService } from '../../services/article.service';
 //Para coger el parámetro de la URL
 import { Router, ActivatedRoute, Params } from '@angular/router';
@@ -36,7 +37,7 @@ export class ArticleComponent implements OnInit {
           if (response.article) {
             console.log(response.article);
             this.article = response.article;
-          }else{
+          } else {
             this._router.navigate(['/404']);
           }
         },
@@ -50,16 +51,54 @@ export class ArticleComponent implements OnInit {
 
   }
 
-  delete(id){
-    this._articleService.delete(id).subscribe(
-      response => {
-        this._router.navigate(['/blog']);
-      },
-      error => {
-        console.log(error);
-        this._router.navigate(['/blog']);
-      }
-    );
+  delete(id) {
+
+    //ALERTA
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "Si aceptas se eliminará para siempre.",
+      icon: 'warning',
+      showCancelButton: true,
+      showCloseButton: true,
+      confirmButtonColor: '#179613',
+      cancelButtonColor: '#C91515',
+      confirmButtonText: '¡Eliminalo!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //Borramos
+        this._articleService.delete(id).subscribe(
+          response => {
+            Swal.fire(
+              '!Artículo Eliminado¡',
+              'El artículo ha sido eliminado correctamente.',
+              'success'
+            ).then((result2) => {
+              if (result2.isConfirmed) {
+                this._router.navigate(['/blog']);
+              }
+            });
+          },
+          error => {
+            console.log(error);
+            //ALERTA
+            Swal.fire({
+              icon: 'error',
+              title: 'Fallo en la Eliminación',
+              text: 'Error al borrar el artículo: ' + error
+            }).then((result) => {
+              if (result.isConfirmed) {
+                this._router.navigate(['/blog']);
+              }
+            });
+            
+          }
+        );
+
+      }//end if
+    });
+
+
   }
 
 }
